implement return random vnf in case removevnf elasticity action is generated
