+++ Dockerfile
@@ -17,7 +17,7 @@ ENV MONGO_DB son-mongo
 #ENV HOST_BROKER int-sp-ath.5gtango.eu
 #ENV CATALOGUE int-sp-ath.5gtango.eu:4011
 #ENV MONITORING_MANAGER int-sp-ath.5gtango.eu:8000
-#ENV REPO qual-sp-bcn.5gtango.eu:4012
+#ENV REPO int-sp-ath.5gtango.eu:4012
 
 ENV HOST_BROKER son-broker
 ENV CATALOGUE tng-cat:4011
