+++ src/main/resources/application-development.properties
@@ -36,4 +36,4 @@ github.repo = https://github.com/sonata-nfv/tng-policy-mngr
 
 monitoring.manager = int-sp-ath.5gtango.eu:8000
 
-tng.rep = qual-sp-bcn.5gtango.eu:4012
+tng.rep = int-sp-ath.5gtango.eu:4012
